client/requirements.txt
mss
numpy
Pillow
PyQt5
pyinstaller